refactor(pto): make PlanMemory the sole local-memory planner
Delete the legacy planner (PTOPlanMemory / OptMemPlanForPipeline) and
promote the largest-first-fit four-gate planner to the canonical
PlanMemory pass, dropping all "modern" naming. Remove the
--plan-memory-impl selector flag and default --plan-memory-order-by-size
to true (the planner's established behavior).

Update lit tests for the now-default planner: the four-gate planner
coalesces tiles more aggressively and relies on InsertSync to guard the
resulting aliasing, so several checks move from separate-address
expectations to the equivalent flag/barrier sequences. Rename the former
plan_memory_modern_* and plan_memory_legacy_* test files accordingly and
drop the obsolete --plan-memory-impl error test.
