+++ test/lit/pto/plan_memory_loop_backedge_liveness.pto
@@ -6,7 +6,7 @@
 // INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 // See LICENSE in the root of the software repository for the full text of the License.
 
-// RUN: ptoas --pto-level=level2 --pto-arch=a3 --plan-memory-impl=modern \
+// RUN: ptoas --pto-level=level2 --pto-arch=a3 \
 // RUN:   --emit-pto-ir --mlir-print-ir-after=pto-plan-memory %s 2>&1 \
 // RUN:   | FileCheck %s
