--- test/lit/pto/plan_memory_modern_multiblock_invalid.pto
@@ -1,22 +1,0 @@
-// Copyright (c) 2026 Huawei Technologies Co., Ltd.
-// This program is free software, you can redistribute it and/or modify it under the terms and conditions of
-// CANN Open Software License Agreement Version 2.0 (the "License").
-// Please refer to the License for details. You may not use this file except in compliance with the License.
-// THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
-// INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
-// See LICENSE in the root of the software repository for the full text of the License.
-
-// RUN: ptoas --pto-level=level2 --pto-arch=a3 \
-// RUN:   --plan-memory-impl=modern --emit-pto-ir %s 2>&1 | FileCheck %s
-
-module {
-  func.func @multiblock_cfg() {
-    cf.br ^next
-  ^next:
-    return
-  }
-}
-
-// CHECK-LABEL: func.func @multiblock_cfg
-// CHECK: cf.br ^bb1
-// CHECK: ^bb1:
