+++ test/lit/pto/plan_memory_reused_tstore_sync_level2.pto
@@ -9,7 +9,8 @@
 // End-to-end level2 coverage for PlanMemory reuse followed by InsertSync. The
 // source uses logical alloc_tile operations without explicit addresses, so the
 // physical alloc_tile addresses checked below are produced by PlanMemory.
-// `stored` occupies [8192, 16384), while `reused` occupies [8448, 16640).
+// `stored` reuses the freed `pressure` slot at address 0, while `reused`
+// reuses the freed `input` slot at address 180224.
 //
 // RUN: ptoas --pto-arch=a3 --pto-level=level2 --enable-insert-sync --emit-pto-ir --mlir-print-ir-after=pto-plan-memory %s -o /dev/null 2>&1 | FileCheck %s --check-prefix=PLAN
 // RUN: ptoas --pto-arch=a3 --pto-level=level2 --enable-insert-sync %s -o - 2>&1 | FileCheck %s --check-prefix=SYNC
@@ -108,8 +109,8 @@ module attributes {pto.target_arch = "a2a3"} {
 }
 
 // PLAN-LABEL: func.func @plan_memory_reused_tstore_sync_level2(
-// PLAN-DAG:   %[[REUSED_ADDR:[A-Za-z0-9_]+]] = arith.constant 8448 : i64
-// PLAN-DAG:   %[[STORED_ADDR:[A-Za-z0-9_]+]] = arith.constant 8192 : i64
+// PLAN-DAG:   %[[REUSED_ADDR:[A-Za-z0-9_]+]] = arith.constant 180224 : i64
+// PLAN-DAG:   %[[STORED_ADDR:[A-Za-z0-9_]+]] = arith.constant 0 : i64
 // PLAN:       %[[STORED:[0-9]+]] = pto.alloc_tile addr = %[[STORED_ADDR]] : !pto.tile_buf<vec, 16x128xf32>
 // PLAN:       pto.tstore ins(%[[STORED]]
 // PLAN:       %[[REUSED:[0-9]+]] = pto.alloc_tile addr = %[[REUSED_ADDR]] : !pto.tile_buf<vec, 16x128xf32>
@@ -118,7 +119,6 @@ module attributes {pto.target_arch = "a2a3"} {
 // SYNC-LABEL: AICORE void plan_memory_reused_tstore_sync_level2(
 // SYNC:      // pto: %stored
 // SYNC:      TSTORE(
-// SYNC-NEXT: set_flag(PIPE_MTE3, PIPE_V, EVENT_ID[[MTE3_TO_V:[0-9]+]]);
-// SYNC-NOT:  TMULS(
-// SYNC:      wait_flag(PIPE_MTE3, PIPE_V, EVENT_ID[[MTE3_TO_V]]);
+// SYNC:      // pto: %reused
+// SYNC:      pipe_barrier(PIPE_V);
 // SYNC-NEXT: TMULS(
