+++ test/lit/pto/plan_memory_spec_level0_reuse.pto
@@ -6,7 +6,7 @@
 // INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 // See LICENSE in the root of the software repository for the full text of the License.
 
-// RUN: ptoas --pto-level=level2 --pto-arch=a3 --plan-memory-impl=modern --emit-pto-ir \
+// RUN: ptoas --pto-level=level2 --pto-arch=a3 --emit-pto-ir \
 // RUN:   --mlir-print-ir-after=pto-plan-memory %s 2>&1 1>/dev/null | FileCheck %s
 
 module attributes {"pto.target_arch" = "a3"} {
