+++ test/lit/pto/tfillpad_plan_memory_inference.pto
@@ -6,7 +6,7 @@
 // INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 // See LICENSE in the root of the software repository for the full text of the License.
 
-// RUN: ptoas --plan-memory-impl=modern %s | FileCheck %s
+// RUN: ptoas %s | FileCheck %s
 
 module {
   func.func @tfillpad_plan_memory_alias() {
