+++ tools/ptoas/ptoas_internal.h
@@ -68,7 +68,6 @@ extern llvm::cl::opt<bool> enableVexpdifFusion;
 extern llvm::cl::opt<bool> enableShapeInference;
 extern llvm::cl::opt<bool> enableVfSimCostmodelOptimization;
 extern llvm::cl::opt<bool> dumpVfSimUnrollTest;
-extern llvm::cl::opt<std::string> planMemoryImpl;
 extern llvm::cl::opt<bool> planMemoryOrderBySize;
 extern llvm::cl::opt<std::string> ptoBuildLevel;
 extern llvm::cl::opt<bool> disableInferLayout;
